requirements.txt
# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
aiofiles>=23.0.0

# Image Processing
opencv-python>=4.8.0

# Data Loading
kaggle>=1.5.0
requests>=2.30.0

# Load Testing
locust>=2.15.0

# Async Support
aiohttp>=3.8.0